+++ components-api/src/main/java/pl/touk/nussknacker/engine/api/ParameterCategoryType.java
@@ -1,0 +1,6 @@
+package pl.touk.nussknacker.engine.api;
+
+public enum ParameterCategoryType {
+    STANDARD,
+    ADVANCED,
+}