+++ designer/client/src/actions/nk/editNode.ts
@@ -1,13 +1,12 @@
-import { getEdgesForNode } from "../../components/graph/node-modal/node/NodeDetails";
-
+import { getEdgesForNode } from "../../components/graph/node-modal/node/useNodeState";
 import { replaceNodeData } from "../../components/graph/node-modal/NodeSwitcherUtils";
-import { Scenario } from "../../components/Process/types";
+import type { Scenario } from "../../components/Process/types";
 import HttpService from "../../http/HttpService";
 import { updateAfterNodeDelete } from "../../reducers/graph/utils";
 import { getGraph } from "../../reducers/selectors/graph";
 import { getProcessDefinitionData } from "../../reducers/selectors/processDefinitionData";
-import { Edge, NodeType, ScenarioGraph, ValidationResult } from "../../types";
-import { ThunkAction } from "../reduxTypes";
+import type { Edge, NodeType, ScenarioGraph, ValidationResult } from "../../types";
+import type { ThunkAction } from "../reduxTypes";
 import { calculateProcessAfterChange } from "./calculateProcessAfterChange";
 import { clearProcessCounts } from "./displayProcessCounts";
 
@@ -32,8 +31,8 @@ export function editScenarioLabels(scenarioLabels: string[]) {
 
 export function editNode(scenarioBefore: Scenario, before: NodeType, after: NodeType, outputEdges?: Edge[]): ThunkAction {
     return async (dispatch) => {
-        const { processName, scenarioGraph } = await dispatch(calculateProcessAfterChange(scenarioBefore, before, after, outputEdges));
-        const response = await HttpService.validateProcess(scenarioBefore.name, processName, scenarioGraph);
+        const scenarioGraph = await dispatch(calculateProcessAfterChange(scenarioBefore, before, after, outputEdges));
+        const response = await HttpService.validateProcess(scenarioBefore.name, scenarioBefore.name, scenarioGraph);
 
         dispatch(clearProcessCounts());
         dispatch({