+++ scenario-compiler/src/main/scala/pl/touk/nussknacker/engine/language/json/JsonParser.scala
@@ -7,7 +7,8 @@ import pl.touk.nussknacker.engine.api.Context
 import pl.touk.nussknacker.engine.api.context.ValidationContext
 import pl.touk.nussknacker.engine.api.expression.ExpressionTypingInfo
 import pl.touk.nussknacker.engine.api.generics.ExpressionParseError
-import pl.touk.nussknacker.engine.api.typed.typing.{TypingResult, Unknown}
+import pl.touk.nussknacker.engine.api.json.decoders.FromJsonSimpleDecoder
+import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypedObjectTypingResult, TypingResult, Unknown}
 import pl.touk.nussknacker.engine.expression.parse.{CompiledExpression, ExpressionParser, TypedExpression}
 import pl.touk.nussknacker.engine.graph.expression.Expression
 import pl.touk.nussknacker.engine.graph.expression.Expression.Language
@@ -25,7 +26,7 @@ object JsonParser extends ExpressionParser {
     parseJson(jsonString).map { json =>
       TypedExpression(
         CompiledJsonExpression(jsonString, json),
-        JsonExpressionTypingInfo
+        JsonExpressionTypingInfo(json)
       )
     }
   }
@@ -54,8 +55,10 @@ object JsonParser extends ExpressionParser {
 
 }
 
-case object JsonExpressionTypingInfo extends ExpressionTypingInfo {
+case class JsonExpressionTypingInfo(json: Json) extends ExpressionTypingInfo {
 
-  // TODO: Right now we just use Unknown but we should create appropriate typing for Jsons in the future
-  override def typingResult: TypingResult = Unknown
+  override val typingResult: TypingResult = {
+    Typed.fromInstance(FromJsonSimpleDecoder.jsonToAny(json))
+  }
+
 }